+++ rkaiq/include/iq_parser_v2/ynr_head_v22.h
@@ -55,13 +55,13 @@ typedef struct CalibDbV2_YnrV22_Calib_s {
 } CalibDbV2_YnrV22_Calib_t;
 
 typedef struct CalibDbV2_YnrV22_LumaParam_s {
-    // M4_ARRAY_DESC("lo_lumaPoint", "f32", M4_SIZE(1,6), M4_RANGE(0,256), "[0,32,64,128,192,256]", M4_DIGIT(0), M4_DYNAMIC(0))
-    float lo_lumaPoint[6];
-    // M4_ARRAY_DESC("lo_ratio", "f32", M4_SIZE(1,6), M4_RANGE(0,2), "[1,1,1,1,1,1]", M4_DIGIT(2), M4_DYNAMIC(0))
+    // M4_ARRAY_DESC("lo_lumaPoint", "u16", M4_SIZE(1,6), M4_RANGE(0,256), "[0,32,64,128,192,256]", M4_DIGIT(0), M4_DYNAMIC(0))
+    uint16_t lo_lumaPoint[6];
+    // M4_ARRAY_DESC("lo_ratio", "f32", M4_SIZE(1,6), M4_RANGE(0,2.0), "[1.0,1.0,1.0,1.0,1.0,1.0]", M4_DIGIT(2), M4_DYNAMIC(0))
     float lo_ratio[6];
-    // M4_ARRAY_DESC("hi_lumaPoint", "f32", M4_SIZE(1,6), M4_RANGE(0,256), "[0,32,64,128,192,256]", M4_DIGIT(0), M4_DYNAMIC(0))
-    float hi_lumaPoint[6];
-    // M4_ARRAY_DESC("hi_ratio", "f32", M4_SIZE(1,6), M4_RANGE(0,2), "[1,1,1,1,1,1]", M4_DIGIT(2), M4_DYNAMIC(0))
+    // M4_ARRAY_DESC("hi_lumaPoint", "u16", M4_SIZE(1,6), M4_RANGE(0,256), "[0,32,64,128,192,256]", M4_DIGIT(0), M4_DYNAMIC(0))
+    uint16_t hi_lumaPoint[6];
+    // M4_ARRAY_DESC("hi_ratio", "f32", M4_SIZE(1,6), M4_RANGE(0,2.0), "[1.0,1.0,1.0,1.0,1.0,1.0]", M4_DIGIT(2), M4_DYNAMIC(0))
     float hi_ratio[6];
 } CalibDbV2_YnrV22_LumaParam_t;
 
@@ -113,9 +113,9 @@ typedef struct CalibDbV2_YnrV22_T_ISO_s {
     float low_dist_adj;
     // M4_NUMBER_DESC("low_weight", "f32", M4_RANGE(0.0, 1.0), "0.5", M4_DIGIT(2))
     float low_weight;
-    // M4_NUMBER_DESC("low_filt1_strength", "s32", M4_RANGE(0.0, 1.0), "0.7", M4_DIGIT(2))
+    // M4_NUMBER_DESC("low_filt1_strength", "f32", M4_RANGE(0.0, 1.0), "0.7", M4_DIGIT(2))
     float low_filt1_strength;
-    // M4_NUMBER_DESC("low_filt2_strength", "s32", M4_RANGE(0.0, 1.0), "0.85", M4_DIGIT(2))
+    // M4_NUMBER_DESC("low_filt2_strength", "f32", M4_RANGE(0.0, 1.0), "0.85", M4_DIGIT(2))
     float low_filt2_strength;
     // M4_NUMBER_DESC("low_bi_weight", "f32", M4_RANGE(0.0, 1.0), "0.2", M4_DIGIT(2))
     float low_bi_weight;
@@ -134,18 +134,18 @@ typedef struct CalibDbV2_YnrV22_T_ISO_s {
     // M4_NUMBER_DESC("hi_nr_weight", "f32", M4_RANGE(0.0, 1.0), "0.78", M4_DIGIT(2))
     float hi_nr_weight;
 
-    // M4_NUMBER_DESC("hi_filter_coeff1_1", "s32", M4_RANGE(0, 15), "7", M4_DIGIT(0))
-    int hi_filter_coeff1_1;
-    // M4_NUMBER_DESC("hi_filter_coeff1_2", "s32", M4_RANGE(0, 15), "6", M4_DIGIT(0))
-    int hi_filter_coeff1_2;
-    // M4_NUMBER_DESC("hi_filter_coeff1_3", "s32", M4_RANGE(0, 15), "3", M4_DIGIT(0))
-    int hi_filter_coeff1_3;
-    // M4_NUMBER_DESC("hi_filter_coeff2_1", "s32", M4_RANGE(0, 15), "6", M4_DIGIT(0))
-    int hi_filter_coeff2_1;
-    // M4_NUMBER_DESC("hi_filter_coeff2_2", "s32", M4_RANGE(0, 15), "5", M4_DIGIT(0))
-    int hi_filter_coeff2_2;
-    // M4_NUMBER_DESC("hi_filter_coeff2_3", "s32", M4_RANGE(0, 15), "3", M4_DIGIT(0))
-    int hi_filter_coeff2_3;
+    // M4_NUMBER_DESC("hi_filter_coeff1_1", "u8", M4_RANGE(0, 15), "7", M4_DIGIT(0))
+    uint8_t hi_filter_coeff1_1;
+    // M4_NUMBER_DESC("hi_filter_coeff1_2", "u8", M4_RANGE(0, 15), "6", M4_DIGIT(0))
+    uint8_t hi_filter_coeff1_2;
+    // M4_NUMBER_DESC("hi_filter_coeff1_3", "u8", M4_RANGE(0, 15), "3", M4_DIGIT(0))
+    uint8_t hi_filter_coeff1_3;
+    // M4_NUMBER_DESC("hi_filter_coeff2_1", "u8", M4_RANGE(0, 15), "6", M4_DIGIT(0))
+    uint8_t hi_filter_coeff2_1;
+    // M4_NUMBER_DESC("hi_filter_coeff2_2", "u8", M4_RANGE(0, 15), "5", M4_DIGIT(0))
+    uint8_t hi_filter_coeff2_2;
+    // M4_NUMBER_DESC("hi_filter_coeff2_3", "u8", M4_RANGE(0, 15), "3", M4_DIGIT(0))
+    uint8_t hi_filter_coeff2_3;
 
     // M4_ARRAY_DESC("rnr_strength", "f32", M4_SIZE(1,17), M4_RANGE(0,16), "1.0", M4_DIGIT(3), M4_DYNAMIC(0))
     float rnr_strength[17];
